+++ libs/slam/src/slam/CRangeBearingKFSLAM2D.cpp
@@ -932,7 +932,7 @@ void  CRangeBearingKFSLAM2D::getAs3DObject( mrpt::opengl::CSetOfObjectsPtr	&outO
 
 		opengl::CEllipsoidPtr ellip = opengl::CEllipsoid::Create();
 
-		ellip->setName( format( "%"PRIuPTR,i ) );
+		ellip->setName( format( "%u",static_cast<unsigned int>(i) ) );
 		ellip->enableShowName(true);
 		ellip->setPose( pointGauss.mean );
 		ellip->setCovMatrix( pointGauss.cov );
